Save and reset canvas color after screenshot
Canvas doesn't have a color style, which is why I originally just removed the property after the screenshot.
However, if we add a color style to the canvas in the future, removing it after the screenshot would be undesirable.

So now canvas's color is saved and reset after the screenshot, even though the current value is just null.