Map routing and indexing (#168)
Added indexes to map (was retroactively taken out from the repo after the new merge)
Added routing to map
Uses Directions API and not the Static Tile API, which allows for 100,000 requests for free https://www.mapbox.com/pricing/#navigation so it should not be a problem. The fetches are cached so it should not use many API requests.
This is only used to give rough estimates for a schedule. It is not suppose to show exactly where the classes are. The direction API are given coordinates that were already included in AntAlmanac's json file.