Add backend instructions to readme, fix potential recursive install bug
Added a section to the README explaining how to get the backend up and running.

Fixed a recursive install bug that happens when the backend submodule was not cloned.
If the submodule didn't exist, `npm run server-install` would actually just call `npm install` in the root directory, which calls server-install again, resulting in an infite loop.