Search Requirements (#317)
* Search Requirements

this.props.toggleSearch() checks if GE, Department, Course Code/Range, or Instructor are still set to their default values. If none of them have been changed, the search is not performed and an error message is displayed. If one of them has been changed, the search is performed as normal.

* Apply suggestions from code review

Added suggested changes from ChaseC99

Co-authored-by: Chase Carnaroli <chasecarnaroli@gmail.com>