genericize SearchForm
SearchForm now takes two props, changeState and fieldName.
changeState is a function that takes two arguments, a field name and a
value, and changes the state of the other component so that the
corresponding fieldName in that component matches its internal state.

The TermSelector invocation in SearchForm has also been updated to
match its pre-refactor functionality.