Staging Action Improvements (#433)
* Add action trigger on label/unlabel

This makes it so adding the "no deploy" label will destroy the stack if it is previously destroyed and deploys the stack if the "no deploy" label is removed.

* Change from npx cdk to aws to destroy stack

Instead of building the cdk and using npx cdk to destroy the stack we can just use the aws command which simplifies many of the build steps.

* Replace Github deployment with URL comment

Instead of dealing with the permission difficulties that come with Github deployments we instead will just use an action that comments the staging URL, and recreates on update.

* Hard code AWS credentials

Instead of using OIDC to provide temporary credentials we will use AWS credentials stored in the repo secrets. With OIDC, there will not be a good way to get AWS credentials when pull requests come from outside the repo.

* Add concurrency and wait for stack status

Because there is no reason for multiple instances of the action to run at once we add concurrency to force only one instance at a time. That being said, when cancelling another workflow instance sometimes the stack may not have finished updating so adding stack status checks ensures the action will wait until the stack is done updating before attempting more changes.