feat(download.ts): add location attribute to custom events
I noticed a TODO asking to implement a location for custom events
once #249 was finished. I noticed that the issue was completed so
I went ahead and implemented it.

Though... it looks a bit funky because CustomEvent.building saves
an ID instead of the actual building name. This happens in
CustomEventDialog.tsx because it needs a key for the list that's
needed in BuildingSelect. Though I noticed that BuildingSelect
gets the name of the building using the format I used here.